Moves simulated test project to Tests/.
