--- LoRaEngine/test/LoRaWan.SimulatedTest/.editorconfig
@@ -1,6 +1,0 @@
-[*.cs]
-
-dotnet_diagnostic.CA1707.severity=suggestion
-
-# CA5394: Do not use insecure randomness
-dotnet_diagnostic.CA5394.severity = none
