+++ Tests/Simulation/IntegrationTestBaseSim.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace LoRaWan.SimulatedTest
+namespace LoRaWan.Tests.Simulation
 {
     using LoRaWan.Tests.Shared;
     using Xunit;
